Split private logic into `GradesServiceInternal`.
